hw5

+++ src/services/contacts.js
@@ -1,13 +1,15 @@
 import { ContactsCollection } from '../db/models/contact.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 
-export const getAllContacts = async ({ page, perPage, sortOrder, sortBy }) => {
+export const getAllContacts = async ({ page, perPage, sortOrder, sortBy, userId }) => {
     const limit = perPage;
     const skip = page > 0 ? (page - 1) * perPage : 0;
 
+    const contactsQuery = ContactsCollection.find({ userId });
+
     const [total, contacts] = await Promise.all([
-        ContactsCollection.countDocuments(),
-        ContactsCollection.find()
+        ContactsCollection.countDocuments(contactsQuery),
+        contactsQuery
             .sort({ [sortBy]: sortOrder })
             .skip(skip)
             .limit(limit),
@@ -21,8 +23,8 @@ export const getAllContacts = async ({ page, perPage, sortOrder, sortBy }) => {
     };
 };
 
-export const getContactById = async (contactId) => {
-    const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+    const contact = await ContactsCollection.findOne({ _id: contactId, userId });
     return contact;
 };
 
@@ -31,14 +33,14 @@ export const createContact = async (newContact) => {
     return contact;
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId, userId });
     return contact;
 };
 
-export const updateContact = async (contactId, updateContact, options = {}) => {
+export const updateContact = async (contactId, updateContact, userId, options = {}) => {
     const contact = await ContactsCollection.findOneAndUpdate(
-        { _id: contactId },
+        { _id: contactId, userId },
         updateContact,
         { new: true, ...options },
     );
